// packages/trading-core/src/price-service/enigma/types/generated/enigma-price-service.ts
export enum DataSourceDTO {
  DEXSCREENER = "DEXSCREENER",
  MANUAL = "MANUAL",
}
export interface ValidationError {
  loc: (string | number)[];
  msg: string;
  type: string;
}

export interface HTTPValidationError {
  detail?: ValidationError[];
}

export interface LiquidityDTO {
  usd: number;
}

export interface LiquidityData {
  name: string;
  usd: number;
  base: number;
  quote: number;
  time: number;
}

export interface TokenRefDTO {
  address: string;
  name: string;
  symbol: string;
}

export interface PriceChangeDTO {
  h24: number | null;
}

export interface MetadataResponse {
  name: string;
  chain_id: string;
  dex_id: string;
  pair_address: string;
  base_token: TokenRefDTO;
  price_native: string;
  price_usd: string;
  decimal: number;
  liquidity?: LiquidityDTO | null;
  price_change: PriceChangeDTO;
  market_cap?: number | null;
  pair_created_at: number;
  updated_at: number;
  source: DataSourceDTO;
}